refactor: Remove UsersTable component and integrate its functionality directly into the Users page
- Deleted the UsersTable component and its associated CSS.
- Integrated the table rendering logic directly into the Users page, utilizing Ant Design's Table component.
- Updated search handling to use requestParams for better state management.
- Added a new MembersTimeSheet component for reporting, including a JSON data file and chart implementation.
