--- src/components/admin-center/users/users-table/users-table.css
@@ -1,3 +1,0 @@
-.users-table-row :hover .email-hover {
-  color: #0d6efd;
-}
